delete: chat 폴더 및 미사용 코드
