--- lib/chat/model/chatting_state.dart
@@ -1,1 +1,0 @@
-enum ChattingState { initial, listView, detailView, reponsed }
