--- lib/chat/model/room/chat_log.dart
@@ -1,28 +1,0 @@
-class ChatLog {
-  final String content;
-  final String sender;
-
-  ChatLog({
-    required this.content,
-    required this.sender,
-  });
-
-  Map<String, dynamic> toJson() => {
-        'content': content,
-        'sender': sender,
-      };
-
-  @override
-  String toString() => 'ChatLog(content: $content, sender: $sender)';
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      (other is ChatLog &&
-          runtimeType == other.runtimeType &&
-          content == other.content &&
-          sender == other.sender);
-
-  @override
-  int get hashCode => content.hashCode ^ sender.hashCode;
-}
