+++ lib/injection.config.dart
@@ -50,6 +50,7 @@ import 'new_chat/infrastructure/repository/chat_repository.dart' as _i605;
 import 'new_chat/infrastructure/repository/chat_rooms_info_repository.dart'
     as _i779;
 import 'route/application/route_bloc.dart' as _i1045;
+import 'route/domain/route_service.dart' as _i807;
 import 'toast/toast_bloc.dart' as _i301;
 
 extension GetItInjectableX on _i174.GetIt {
@@ -73,6 +74,7 @@ extension GetItInjectableX on _i174.GetIt {
     gh.factory<_i1045.RouteBloc>(() => _i1045.RouteBloc());
     gh.singleton<_i301.ToastBloc>(() => _i301.ToastBloc());
     gh.lazySingleton<List<String>>(() => homeRegisterModule.texts);
+    gh.lazySingleton<_i807.RouteService>(() => _i807.RouteService());
     gh.lazySingleton<_i141.IAuthExternalProvider>(
         () => _i914.AuthExternalProvider());
     gh.lazySingleton<_i690.IAuthLocalProvider>(
@@ -102,14 +104,16 @@ extension GetItInjectableX on _i174.GetIt {
           authInternalProvider: gh<_i5.IAuthInternalProvider>(),
           authLocalProvider: gh<_i690.IAuthLocalProvider>(),
         ));
-    gh.factory<_i810.ChatDrawerBloc>(
-        () => _i810.ChatDrawerBloc(gh<_i779.ChatRoomsInfoRepository>()));
     gh.lazySingleton<_i665.IMypageInternalProvider>(
         () => _i487.MypageInternalProvider(gh<_i1053.FetchService>()));
     gh.factory<_i659.ChatSessionBloc>(() =>
         _i659.ChatSessionBloc(chatRepository: gh<_i605.ChatRepository>()));
     gh.lazySingleton<_i392.IMypageRepository>(() => _i936.MypageRepository(
         mypageProvider: gh<_i665.IMypageInternalProvider>()));
+    gh.factory<_i810.ChatDrawerBloc>(() => _i810.ChatDrawerBloc(
+          gh<_i779.ChatRoomsInfoRepository>(),
+          gh<_i659.ChatSessionBloc>(),
+        ));
     gh.factory<_i317.LoginBloc>(
         () => _i317.LoginBloc(authRepository: gh<_i667.IAuthRepository>()));
     gh.factory<_i835.EditProfileBloc>(() =>
