--- lib/personal/model/user.dart
@@ -1,11 +1,0 @@
-class User {
-  final String imagePath;
-  final String name;
-  final String id;
-
-  const User({
-    required this.imagePath,
-    required this.name,
-    required this.id,
-  });
-}
