--- lib/personal/view/components/button_widget.dart
@@ -1,23 +1,0 @@
-import 'package:flutter/material.dart';
-
-class ButtonWidget extends StatelessWidget {
-  final String text;
-  final VoidCallback onClicked;
-
-  const ButtonWidget({
-    super.key,
-    required this.text,
-    required this.onClicked,
-  });
-
-  @override
-  Widget build(BuildContext context) => ElevatedButton(
-        style: ElevatedButton.styleFrom(
-          foregroundColor: Colors.white,
-          shape: const StadiumBorder(),
-          padding: const EdgeInsets.symmetric(horizontal: 32, vertical: 12),
-        ),
-        onPressed: onClicked,
-        child: Text(text),
-      );
-}
